Move component into another folder
