+++ ui/components/app/currency-input/index.scss
