+++ ui/components/ui/ui-components.scss
@@ -15,7 +15,6 @@
 @import 'color-indicator/color-indicator';
 @import 'confusable/index';
 @import 'currency-display/index';
-@import 'currency-input/index';
 @import 'definition-list/definition-list';
 @import 'dialog/dialog';
 @import 'dropdown/dropdown';
